requirements.txt
pillow==11.1.0
pytesseract==0.3.13
python-docx==1.1.2
streamlit==1.41.1
et_xmlfile==2.0.0